+++ blocks/detail-viewer/src/components/ObjectRenderer/DefaultRenderer/index.jsx
@@ -1,5 +1,0 @@
-import { withBlock } from '@appsemble/react';
-
-import DefaultRenderer from './DefaultRenderer';
-
-export default withBlock(DefaultRenderer);