+++ CHANGELOG.md
@@ -8,6 +8,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 - App: Add support for app description field.
+- Block: Rewrite and simplify detail-viewer block
 - Editor: Add support for app description field.
 - Server: Add support for app description field.